Export static function as `x509_import_spki()`
